🐛 Corrections de multiples bugs empêchant l'ajout et l'édition des entités

+++ src/Entity/Booking.php
@@ -27,7 +27,7 @@ class Booking
     private ?\DateTimeInterface $end_date = null;
 
     #[ORM\Column]
-    private ?float $amount = null;
+    private ?int $amount = null;
 
     #[ORM\Column]
     private ?bool $status = null;
@@ -35,7 +35,7 @@ class Booking
     #[ORM\Column(type: Types::DATETIME_MUTABLE)]
     private ?\DateTimeInterface $created_at = null;
 
-    #[ORM\Column(type: Types::DATETIME_MUTABLE)]
+    #[ORM\Column(type: Types::DATETIME_MUTABLE, nullable: false)]
     private ?\DateTimeInterface $updated_at = null;
 
     #[ORM\ManyToOne(inversedBy: 'bookings')]
@@ -101,12 +101,12 @@ class Booking
         return $this;
     }
 
-    public function getAmount(): ?float
+    public function getAmount(): ?int
     {
         return $this->amount;
     }
 
-    public function setAmount(float $amount): static
+    public function setAmount(int $amount): static
     {
         $this->amount = $amount;
